Configure listeners by default in application, ws ,ws-cat; add client timeout configuration to Remoting
