+++ oap-application/src/main/resources/META-INF/oap-module.conf
@@ -10,5 +10,18 @@
       }
       supervision.supervise = true
     }
+    secure-http-listener {
+      dependsOn = [oap-http-server]
+      implementation = oap.http.SecureHttpListener
+      parameters {
+        port = 8980
+        server = "@service:oap-http-server"
+        keystoreLocation = ""
+        keystorePassword = ""
+      }
+      supervision {
+        thread = true
+      }
+    }
   }
 }
