+++ oap-application/src/test/resources/oap/application/application.conf
@@ -6,4 +6,7 @@
   }
 
   so.i2 = "100"
+
+  secure-http-listener.parameters.keystoreLocation = "src/test/resources/oap/application/server_keystore.jks"
+  secure-http-listener.parameters.keystorePassword = "123456"
 }
