+++ oap-application/src/test/resources/oap/application/modules/m1.conf
@@ -15,30 +15,7 @@
       remoteName = "ServiceTwo"
       certificateLocation = "src/test/resources/oap/application/client_truststore.jks"
       certificatePassword = "123456"
+      timeout = 1000
     }
-    plain-http-listener {
-      dependsOn = [oap-http-server]
-      implementation = oap.http.PlainHttpListener
-      parameters {
-        port = 8880
-        server = "@service:oap-http-server"
-      }
-      supervision {
-        thread = true
-      }
-    }
-    secure-http-listener {
-      dependsOn = [oap-http-server]
-      implementation = oap.http.SecureHttpListener
-      parameters {
-        port = 8980
-        server = "@service:oap-http-server"
-        keystoreLocation = "src/test/resources/oap/application/server_keystore.jks"
-        keystorePassword = "123456"
-      }
-      supervision {
-        thread = true
-       }
-     }
   }
 }
