+++ oap-http/pom.xml
@@ -32,7 +32,7 @@
     <parent>
         <artifactId>oap</artifactId>
         <groupId>oap</groupId>
-        <version>1.9.21</version>
+        <version>1.9.22</version>
     </parent>
     <dependencies>
         <dependency>
