+++ oap-logstream/pom.xml
@@ -33,7 +33,7 @@
     <parent>
         <groupId>oap</groupId>
         <artifactId>oap</artifactId>
-        <version>1.9.21</version>
+        <version>1.9.22</version>
     </parent>
 
     <dependencies>
