+++ oap-maven/pom.xml
@@ -34,7 +34,7 @@
     <parent>
         <artifactId>oap</artifactId>
         <groupId>oap</groupId>
-        <version>1.9.21</version>
+        <version>1.9.22</version>
     </parent>
 
     <dependencies>
