+++ oap-ws-cat/src/main/java/oap/ws/cat/CatApi.java
@@ -1,4 +1,4 @@
-/*
+package oap.ws.cat;/*
  * The MIT License (MIT)
  *
  * Copyright (c) Open Application Platform Authors
@@ -21,8 +21,6 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  */
-
-package oap.ws;
 
 import com.google.common.collect.Iterables;
 import oap.http.HttpResponse;
