+++ oap-ws-cat/src/main/resources/META-INF/oap-module.conf
@@ -1,0 +1,17 @@
+{
+  name = oap-ws-cat
+  dependsOn = [oap-http]
+  services {
+    local-http-listener {
+      dependsOn = [oap-http-server]
+      implementation = oap.http.LocalHttpListener
+      parameters {
+        port = 9090
+        server = "@service:oap-http-server"
+      }
+      supervision {
+        thread = true
+      }
+    }
+  }
+}
