+++ oap-ws/src/main/resources/META-INF/oap-module.conf
@@ -11,5 +11,15 @@
         supervise = true
       }
     }
+    plain-http-listener {
+      implementation = oap.http.PlainHttpListener
+      parameters {
+        port = 8880
+        server = "@service:oap-http-server"
+      }
+      supervision {
+        thread = true
+      }
+    }
   }
 }
