+++ pom.xml
@@ -29,7 +29,7 @@
     <packaging>pom</packaging>
     <groupId>oap</groupId>
     <artifactId>oap</artifactId>
-    <version>1.9.21</version>
+    <version>1.9.22</version>
 
     <name>oap</name>
 
@@ -49,6 +49,7 @@
         <module>oap-etl</module>
         <module>oap-logstream</module>
         <module>oap-maven</module>
+        <module>oap-ws-cat</module>
     </modules>
 
     <repositories>
